Add timeout of 10 seconds
This library does not have a timeout and python does not have a default timeout in the CPython implementation, hence if this gets stuck for some reason and in practice it does get stuck often enough... then it hangs there forever waiting. So we need to time it out eventually.